# openapi.yml
openapi: 3.1.0
info:
  title: API de réservation de salles
  version: 1.0.0
  description: |
    Contrat de l'API de gestion des salles, équipements, organisateurs et réservations.
    Les dates et heures utilisent le format ISO 8601 avec fuseau ou décalage UTC.
servers:
  - url: http://localhost:8080
    description: Environnement local
...
paths:
...
  /api/rooms:
    post:
      tags: [Rooms]
      summary: Créer une salle
      description: |
        Crée une salle avec un nom unique, la référence d'un bâtiment existant, un étage
        valide dans ce bâtiment et une capacité strictement positive. Son statut initial
        est `AVAILABLE`. Les codes d'équipement fournis doivent correspondre à des
        équipements existants.
      operationId: createRoom
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRoomRequest'
      responses:
        '201':
          description: Salle créée
          headers:
            Location:
              description: URI de la salle créée
              schema:
                type: string
                format: uri
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoomResponse'
        '400':
          $ref: '#/components/responses/ValidationBadRequest'
        '404':
          $ref: '#/components/responses/RoomCreationResourceNotFound'
        '409':
          $ref: '#/components/responses/ResourceAlreadyExists'
    get:
      tags: [Rooms]
      summary: Lister les salles
      description: Retourne toutes les salles avec leur statut et leurs équipements.
      operationId: listRooms
      responses:
        '200':
          description: Liste des salles, triée par nom
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RoomResponse'
  /api/rooms/available:
    get:
      tags: [Rooms]
      summary: Rechercher les salles disponibles et compatibles
      description: |
        Ne conserve que les salles ayant le statut `AVAILABLE`, une capacité supérieure
        ou égale à `capacity`, tous les équipements demandés et aucune réservation
        confirmée chevauchant la période `[start, end[`.

        Les résultats sont classés par nombre de places inutilisées croissant
        (`capacité - capacity`), puis par nom sans tenir compte de la casse, puis par
        identifiant. Cette opération ne crée aucune réservation et renvoie une liste
        vide lorsqu'aucune salle n'est compatible.
      operationId: findAvailableRooms
      parameters:
        - $ref: '#/components/parameters/Start'
        - $ref: '#/components/parameters/End'
        - name: capacity
          in: query
          required: true
          description: Nombre de participants à accueillir
          schema:
            type: integer
            format: int32
            minimum: 1
          example: 25
        - name: equipment
          in: query
          required: false
          description: Codes des équipements exigés, séparés par des virgules
          style: form
          explode: false
          schema:
            type: array
            uniqueItems: true
            items:
              $ref: '#/components/schemas/EquipmentCode'
          example: [PROJECTOR, VIDEO_CONFERENCE]
      responses:
        '200':
          description: Salles compatibles, éventuellement vide
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AvailableRoomResponse'
              examples:
                noAvailableRoom:
                  summary: Aucune salle ne correspond aux critères
                  value: []
        '400':
          $ref: '#/components/responses/AvailabilityBadRequest'
  /api/rooms/{roomId}:
    parameters:
      - $ref: '#/components/parameters/RoomId'
    get:
      tags: [Rooms]
      summary: Consulter une salle
      description: Retourne la salle identifiée, y compris son statut et ses équipements.
      operationId: getRoom
      responses:
        '200':
          description: Salle trouvée
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoomResponse'
        '404':
          $ref: '#/components/responses/RoomNotFound'
    put:
      tags: [Rooms]
      summary: Remplacer les informations d'une salle
      description: |
        Remplace le nom, la référence du bâtiment, l'étage et la capacité de la salle.
        L'étage doit exister dans le bâtiment référencé. Le statut et les équipements
        ne sont pas modifiés. Le nouveau nom doit rester unique.
      operationId: updateRoom
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateRoomRequest'
      responses:
        '200':
          description: Salle mise à jour
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoomResponse'
        '400':
          $ref: '#/components/responses/ValidationBadRequest'
        '404':
          $ref: '#/components/responses/RoomOrBuildingNotFound'
        '409':
          $ref: '#/components/responses/ResourceAlreadyExists'
  /api/rooms/{roomId}/status:
    parameters:
      - $ref: '#/components/parameters/RoomId'
    patch:
      tags: [Rooms]
      summary: Modifier le statut d'une salle
      description: |
        Place la salle dans l'état `AVAILABLE` ou `MAINTENANCE`. Une salle en
        maintenance n'est plus proposée ni réservable. Les réservations existantes
        sont conservées.
      operationId: updateRoomStatus
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateRoomStatusRequest'
      responses:
        '200':
          description: Statut mis à jour
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoomResponse'
        '400':
          $ref: '#/components/responses/ValidationBadRequest'
        '404':
          $ref: '#/components/responses/RoomNotFound'
  /api/rooms/{roomId}/equipment:
    parameters:
      - $ref: '#/components/parameters/RoomId'
    put:
      tags: [Rooms, Equipment]
      summary: Remplacer les équipements d'une salle
      description: |
        La liste fournie remplace entièrement la liste actuelle. Une liste vide retire
        tous les équipements. Chaque code doit désigner un équipement existant.
      operationId: replaceRoomEquipment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReplaceRoomEquipmentRequest'
      responses:
        '200':
          description: Équipements de la salle mis à jour
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoomResponse'
        '400':
          $ref: '#/components/responses/ValidationBadRequest'
        '404':
          $ref: '#/components/responses/RoomOrEquipmentNotFound'
...
components:
  parameters:
...
    RoomId:
      name: roomId
      in: path
      required: true
      schema:
        $ref: '#/components/schemas/ResourceId'
...
    Start:
      name: start
      in: query
      required: true
      description: Début de la période recherchée
      schema:
        type: string
        format: date-time
      example: '2026-10-15T14:00:00+02:00'
    End:
      name: end
      in: query
      required: true
      description: Fin de la période recherchée
      schema:
        type: string
        format: date-time
      example: '2026-10-15T16:00:00+02:00'
  responses:
    ValidationBadRequest:
      description: Un ou plusieurs champs de la requête sont invalides
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
    AvailabilityBadRequest:
      description: Les critères ou la période de recherche sont invalides
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
          examples:
            invalidPeriod:
              summary: Période invalide
              value:
                code: INVALID_RESERVATION_PERIOD
                message: La date de début doit être antérieure à la date de fin
                timestamp: '2026-10-01T13:45:12Z'
                path: /api/rooms/available
                details: {}
                fieldErrors: {}
            invalidCapacity:
              summary: Capacité invalide
              value:
                code: VALIDATION_ERROR
                message: La requête contient des données invalides
                timestamp: '2026-10-01T13:45:12Z'
                path: /api/rooms/available
                details: {}
                fieldErrors:
                  capacity: doit être strictement positive
...
    RoomCreationResourceNotFound:
      description: Le bâtiment ou au moins un équipement demandé n'existe pas
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
          examples:
            buildingNotFound:
              value:
                code: BUILDING_NOT_FOUND
                message: Le bâtiment 3 n'existe pas
                timestamp: '2026-10-01T13:45:12Z'
                path: /api/rooms
                details:
                  buildingId: 3
                fieldErrors: {}
            equipmentNotFound:
              value:
                code: EQUIPMENT_NOT_FOUND
                message: L'équipement PROJECTOR n'existe pas
                timestamp: '2026-10-01T13:45:12Z'
                path: /api/rooms
                details:
                  equipmentCode: PROJECTOR
                fieldErrors: {}
    RoomOrBuildingNotFound:
      description: La salle ou le bâtiment demandé n'existe pas
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
          examples:
            roomNotFound:
              value:
                code: ROOM_NOT_FOUND
                message: La salle 7 n'existe pas
                timestamp: '2026-10-01T13:45:12Z'
                path: /api/rooms/7
                details:
                  roomId: 7
                fieldErrors: {}
            buildingNotFound:
              value:
                code: BUILDING_NOT_FOUND
                message: Le bâtiment 3 n'existe pas
                timestamp: '2026-10-01T13:45:12Z'
                path: /api/rooms/7
                details:
                  buildingId: 3
                fieldErrors: {}
    RoomNotFound:
      description: La salle demandée n'existe pas
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
          example:
            code: ROOM_NOT_FOUND
            message: La salle 7 n'existe pas
            timestamp: '2026-10-01T13:45:12Z'
            path: /api/rooms/7
            details:
              roomId: 7
            fieldErrors: {}
...
    RoomOrEquipmentNotFound:
      description: La salle ou au moins un équipement demandé n'existe pas
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
          examples:
            roomNotFound:
              value:
                code: ROOM_NOT_FOUND
                message: La salle 7 n'existe pas
                timestamp: '2026-10-01T13:45:12Z'
                path: /api/rooms/7/equipment
                details:
                  roomId: 7
                fieldErrors: {}
            equipmentNotFound:
              value:
                code: EQUIPMENT_NOT_FOUND
                message: L'équipement VIDEO_CONFERENCE n'existe pas
                timestamp: '2026-10-01T13:45:12Z'
                path: /api/rooms/7/equipment
                details:
                  equipmentCode: VIDEO_CONFERENCE
                fieldErrors: {}
...
    ResourceAlreadyExists:
      description: Une valeur soumise à une contrainte d'unicité est déjà utilisée
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
...
  schemas:
    ResourceId:
      type: integer
      format: int64
      minimum: 1
      example: 7
    RoomStatus:
      type: string
      enum: [AVAILABLE, MAINTENANCE]
...
    EquipmentCode:
      type: string
      pattern: '^[A-Z][A-Z0-9_]{1,49}$'
      example: PROJECTOR
    BuildingName:
      type: string
      minLength: 1
      maxLength: 100
      description: Nom unique du bâtiment, sans tenir compte de la casse
      example: Bâtiment A
    FloorNumber:
      type: integer
      format: int32
      minimum: 0
      maximum: 199
      description: Doit être inférieur au nombre d'étages du bâtiment référencé
      example: 2
    NumberOfFloors:
      type: integer
      format: int32
      minimum: 1
      maximum: 200
      description: Nombre d'étages, rez-de-chaussée inclus
      example: 5
...
    BuildingResponse:
      type: object
      additionalProperties: false
      required: [id, name, numberOfFloors]
      properties:
        id:
          $ref: '#/components/schemas/ResourceId'
        name:
          $ref: '#/components/schemas/BuildingName'
        numberOfFloors:
          $ref: '#/components/schemas/NumberOfFloors'
    CreateRoomRequest:
      type: object
      additionalProperties: false
      required: [name, buildingId, floor, capacity]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
          example: Orion
        buildingId:
          $ref: '#/components/schemas/ResourceId'
        floor:
          $ref: '#/components/schemas/FloorNumber'
        capacity:
          type: integer
          format: int32
          minimum: 1
          example: 30
        equipmentCodes:
          type: array
          uniqueItems: true
          default: []
          items:
            $ref: '#/components/schemas/EquipmentCode'
    UpdateRoomRequest:
      type: object
      additionalProperties: false
      required: [name, buildingId, floor, capacity]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
          example: Orion
        buildingId:
          $ref: '#/components/schemas/ResourceId'
        floor:
          $ref: '#/components/schemas/FloorNumber'
        capacity:
          type: integer
          format: int32
          minimum: 1
          example: 32
    UpdateRoomStatusRequest:
      type: object
      additionalProperties: false
      required: [status]
      properties:
        status:
          $ref: '#/components/schemas/RoomStatus'
    ReplaceRoomEquipmentRequest:
      type: object
      additionalProperties: false
      required: [equipmentCodes]
      properties:
        equipmentCodes:
          type: array
          uniqueItems: true
          items:
            $ref: '#/components/schemas/EquipmentCode'
          example: [PROJECTOR, WHITEBOARD]
    RoomSummaryResponse:
      type: object
      required: [id, name, building, floor, capacity, status]
      properties:
        id:
          $ref: '#/components/schemas/ResourceId'
        name:
          type: string
          example: Orion
        building:
          $ref: '#/components/schemas/BuildingResponse'
        floor:
          $ref: '#/components/schemas/FloorNumber'
        capacity:
          type: integer
          format: int32
          minimum: 1
          example: 30
        status:
          $ref: '#/components/schemas/RoomStatus'
    RoomResponse:
      allOf:
        - $ref: '#/components/schemas/RoomSummaryResponse'
        - type: object
          required: [equipment]
          properties:
            equipment:
              type: array
              items:
                $ref: '#/components/schemas/EquipmentResponse'
    AvailableRoomResponse:
      allOf:
        - $ref: '#/components/schemas/RoomResponse'
        - type: object
          required: [unusedCapacity]
          properties:
            unusedCapacity:
              type: integer
              format: int32
              minimum: 0
              description: Capacité de la salle moins la capacité demandée
              example: 5
...
    EquipmentResponse:
      type: object
      additionalProperties: false
      required: [id, code, label]
      properties:
        id:
          $ref: '#/components/schemas/ResourceId'
        code:
          $ref: '#/components/schemas/EquipmentCode'
        label:
          type: string
          example: Vidéoprojecteur
...
    ApiErrorResponse:
      type: object
      description: |
        Format commun à toutes les erreurs de l'API.

        | Code | HTTP | Signification |
        | --- | ---: | --- |
        | `VALIDATION_ERROR` | 400 | Un ou plusieurs champs sont invalides. |
        | `INVALID_RESERVATION_PERIOD` | 400 | La période est passée, inversée ou dépasse huit heures. |
        | `BUILDING_NOT_FOUND` | 404 | Le bâtiment demandé n'existe pas. |
        | `ROOM_NOT_FOUND` | 404 | La salle demandée n'existe pas. |
        | `ORGANIZER_NOT_FOUND` | 404 | L'organisateur demandé n'existe pas. |
        | `RESERVATION_NOT_FOUND` | 404 | La réservation demandée n'existe pas. |
        | `EQUIPMENT_NOT_FOUND` | 404 | Au moins un code d'équipement n'existe pas. |
        | `ROOM_CAPACITY_EXCEEDED` | 409 | La salle choisie est trop petite. |
        | `MISSING_REQUIRED_EQUIPMENT` | 409 | La salle choisie ne possède pas tous les équipements demandés. |
        | `ROOM_ALREADY_RESERVED` | 409 | Une réservation confirmée chevauche la période. |
        | `ROOM_UNAVAILABLE` | 409 | La salle choisie est en maintenance. |
        | `NO_COMPATIBLE_ROOM` | 409 | Aucune salle ne convient à l'attribution automatique. |
        | `RESERVATION_ALREADY_CANCELLED` | 409 | La réservation est déjà annulée. |
        | `RESOURCE_ALREADY_EXISTS` | 409 | Un nom, code ou e-mail unique est déjà utilisé. |
        | `BUILDING_FLOOR_COUNT_CONFLICT` | 409 | Le nombre d'étages est incompatible avec une localisation existante. |
      additionalProperties: false
      required: [code, message, timestamp, path]
      properties:
        code:
          type: string
          enum:
            - VALIDATION_ERROR
            - INVALID_RESERVATION_PERIOD
            - BUILDING_NOT_FOUND
            - ROOM_NOT_FOUND
            - ORGANIZER_NOT_FOUND
            - RESERVATION_NOT_FOUND
            - EQUIPMENT_NOT_FOUND
            - ROOM_CAPACITY_EXCEEDED
            - MISSING_REQUIRED_EQUIPMENT
            - ROOM_ALREADY_RESERVED
            - ROOM_UNAVAILABLE
            - NO_COMPATIBLE_ROOM
            - RESERVATION_ALREADY_CANCELLED
            - RESOURCE_ALREADY_EXISTS
            - BUILDING_FLOOR_COUNT_CONFLICT
        message:
          type: string
        timestamp:
          type: string
          format: date-time
        path:
          type: string
        details:
          type: object
          additionalProperties: true
          default: {}
        fieldErrors:
          type: object
          additionalProperties:
            type: string
          default: {}